requirements.txt
requests
cassandra-driver
pyspark
pandas
scikit-learn
matplotlib
numpy
